Added api.py Added helper.py
Finished to convert code to OO paradigm.
