+++ main.py
@@ -1,19 +1,15 @@
-#~-~-~-~-~-~-~-~-~-~-~-~-~-#
-#          gSync           #
-#    Gabriel Santamaria    #
-#        Apache 2.0        #
-#~-~-~-~-~-~-~-~-~-~-~-~-~-#
+# File : main.py
+# Created by gabys
+# Date : 7/09/2018
+# License : Apache 2.0
 
-import abc
-# Imports
-import datetime
-import json
+# Importing Python libraries
 import os
 import sys
 import urllib.request
-import zipfile
 
-import requests
+# Importing API classes
+from api import *
 
 # CONFIGURATION : Configure your script here
 steam_key = ''  # Your Steam API key
@@ -22,212 +18,58 @@ gmad = ''  # Your path to gmad.exe
 gmpublish = ''  # Your path to gmpublish.exe
 
 
-class Helper:
-
-    @staticmethod
-    def check_times(steam, app_id, file_id, github):
-        return steam.get_last_update(app_id, file_id) >= github.get_last_update()
-
-    @staticmethod
-    def is_zip(file_type):
-        mimes = {
-            'application/zip',
-            'application/octet-stream',
-            'application/x-zip-compressed',
-            'multipart/x-zip'
-        }
-        for v in mimes:
-            if v == file_type:
-                return True
-        return False
-
-    @staticmethod
-    def unzip(path, file_name):
-        if not os.path.isfile(path + '/' + file_name):
-            raise Exception(
-                'A problem occurred while trying to unzip the file archive : file do not exists nor is a valid one.')
-        f = zipfile.ZipFile(path + '/' + file_name, 'r')
-        f.extractall(path)
-        f.close()
-
-    @staticmethod
-    def create_directory(directory_name):
-        if not os.path.exists(directory_name):
-            os.makedirs(directory_name)
-
-    @staticmethod
-    def create_empty_file(file_name, relative_directory):
-        open(relative_directory + '/' + file_name, 'w').close()
-
-    # TOOK & ADAPTED FROM THIS POST :
-    # https://stackoverflow.com/questions/185936/how-to-delete-the-contents-of-a-folder-in-python
-    @staticmethod
-    def clear_directory(folder):
-        for file in os.listdir(folder):
-            path = os.path.join(folder, file)
-            try:
-                _, ext = os.path.splitext(path)
-                if os.path.isfile(path) and ext != '.json':
-                    os.unlink(path)
-            except Exception as e:
-                print(e)
-
-    @staticmethod
-    def get_base_directory():
-        blacklist = {
-            'maps',
-            'backgrounds',
-            'gamemodes',
-            'materials',
-            'lua',
-            'scenes',
-            'models',
-            'scripts',
-            'particles',
-            'sound',
-            'resource'
-        }
-        i = 0
-        directories = [f.path for f in os.scandir('tmp') if f.is_dir()]
-        for p in directories:
-            if os.path.isdir(os.path.join('tmp', p)):
-                i = i + 1
-        if i > 1:
-            return 'tmp'
-        else:
-            if directories[0] in blacklist:
-                return 'tmp'
-            else:
-                return directories[0]
-
-
-class API(metaclass=abc.ABCMeta):
-    api_key: str = ""
-
-    def __init__(self, api_key):
-        self.api_key = api_key
-
-    @abc.abstractmethod
-    def get_last_update(self):
-        """ Gets the last update time """
-        pass
-
-
-class Steam(API):
-
-    def __init__(self, api_key):
-        super().__init__(api_key)
-
-    def get_last_update(self, app_id, file_id):
-        base_url = 'https://api.steampowered.com/ISteamRemoteStorage/GetPublishedFileDetails/v1/'
-        parameters = {
-            'key': self.api_key,
-            'app_id': app_id,
-            'publishedfileids[0]': file_id,
-            'itemcount': 1
-        }
-        r = requests.post(url=base_url, data=parameters)
-        data = json.loads(r.text)
-        try:
-            data['response']['publishedfiledetails'][0]['time_updated']
-        except NameError:
-            raise Exception('Can\'t get the latest update time of the following WS item ' + file_id)
-
-        time = data['response']['publishedfiledetails'][0]['time_updated']
-        return time
-
-
-def createWorkshopArchive(path, basePath):
-    fullPath = os.path.dirname(os.path.realpath(__file__)) + '/' + basePath
-    status = os.system(gmad + ' create -folder ' + fullPath + ' -out ' + fullPath)
+def create_ws_archive(base_path) -> None:
+    """
+    Create a workshop archive (.gma file) in the path "base_path"
+    Returns:
+        None: Nothing
+    """
+    full_path = os.path.dirname(os.path.realpath(__file__)) + '/' + base_path
+    status = os.system(gmad + ' create -folder ' + full_path + ' -out ' + full_path)
     if status != 0:
-        raise Exception('The GMA proccess ended with an error.')
+        raise Exception('The GMA process ended with an error.')
 
 
-def updateWorkshopItem(gmaPath, fileId, changes):
-    fullPath = os.path.dirname(os.path.realpath(__file__)) + '/' + gmaPath
-    status = os.system(gmpublish + ' update -addon "' + fullPath + '" -id "' + fileId + '" -changes "' + changes + '"')
+def update_ws_item(gma_path, file_id, changes) -> None:
+    """
+    Update a workshop item (with id "file_id" using gmpublish.exe
+    Returns:
+        None: Nothing
+    """
+    full_path = os.path.dirname(os.path.realpath(__file__)) + '/' + gma_path
+    status = os.system(
+        gmpublish + ' update -addon "' + full_path + '" -id "' + file_id + '" -changes "' + changes + '"')
     if status != 0:
-        raise Exception('The GMPUBLISH proccess ended with an error.')
+        raise Exception('The GMPUBLISH process ended with an error.')
 
 
-# --------------------#
-#    GITHUB  PART    #
-# --------------------#
+# Main function
+def main():
+    if len(sys.argv) < 3:
+        raise Exception('Missing arguments.')
 
-class Github(API):
-    repository_name: str = ""
-    base_url: str = ""
+    app_id = str(sys.argv[1])  # Steam app ID
+    file_id = str(sys.argv[2])  # Steam Workshop file ID
+    repo_name = str(sys.argv[3])  # Github repository name format : <owner/name>
 
-    def __init__(self, api_key, repository_name):
-        super().__init__(api_key)
-        self.repository_name = repository_name
-        self.base_url = 'https://api.github.com/repos/' + self.repository_name + '/releases/latest?access_token=' + self.api_key
+    # Since we're just supporting Github and Steam, don't need to do any check on which to use
+    steam = Steam(steam_key, app_id, file_id)
+    github = Github(github_key, repo_name)
 
-    def get_last_update(self):
-        r = requests.get(self.base_url)
-        data = json.loads(r.text)
-
-        try:
-            iso = datetime.datetime.strptime(data['published_at'], "%Y-%m-%dT%H:%M:%S%z")
-        except NameError:
-            raise Exception('Can\'t fetch any attached asset to the latest release of ' + self.repository_name)
-
-        return iso.timestamp()
-
-    def get_release_data(self):
-        """
-        Gets the latest release data from the Github website
-        Returns: An array containing the usefull data
-
-        """
-        r = requests.get(self.base_url)
-        data = json.loads(r.text)
-
-        try:
-            download_url = data['assets'][0]['browser_download_url']
-        except NameError:
-            raise Exception('Can\'t fetch any attached asset to the latest release of ' + self.repository_name)
-        try:
-            data['body']
-        except NameError:
-            raise Exception('An error occurred while trying to access to the latest release of ' + self.repository_name)
-
-        if not Helper.is_zip(data['assets'][0]['content_type']):
-            raise Exception('The latest update asset isn\'t a valid ZIP file')
-
-        return {
-            'downloadUrl': download_url,
-            'body': data['body'],
-            'name': data['assets'][0]['name'],
-            'size': data['assets'][0]['size']
-        }
-
-
-# ------------------------------#
-#                              #
-#         MAIN FUNCTION        #
-#                              #
-# ------------------------------#
-def Main():
-    appid = str(sys.argv[1])  # Steam's APPID
-    fileid = str(sys.argv[2])  # Steam's Workshop FILEID
-    repoName = str(sys.argv[3])  # Github's repositery name FORMAT : <owner/name>
-
-    if checkTimes(steam_key, appid, fileid, github_key, repoName):
+    if helper.check_times(steam, github):
         print('Everything is up to date. Your workshop file has already been synchronized with your Github repository.')
         return 0
 
-    data = getGithubReleaseData(github_key, repoName)
+    data = github.get_release_data()
 
-    # Checking /tmp/ directory + creating the empty file
+    # Checking & creating /tmp/ directory + creating the empty file
     print('Preparing the download...')
-    createDirectory('tmp')
-    createEmptyFile(data['name'], 'tmp')
+    helper.create_directory('tmp')
+    helper.create_empty_file(data['name'], 'tmp')
 
     # Downloading and writing the file into the previously created empty file
     t = datetime.datetime.now().timestamp()
-    print('Starting to download the latest release of https://github.com/' + repoName + ' ... ' + str(
+    print('Starting to download the latest release of https://github.com/' + repo_name + ' ... ' + str(
         data['size']) + ' ...')
     urllib.request.urlretrieve(data['downloadUrl'], 'tmp/' + data['name'])
     print('Download finished with success. It took : ' + str(datetime.datetime.now().timestamp() - t) + ' seconds')
@@ -235,30 +77,33 @@ def Main():
     # Unzipping the downloaded file
     t = datetime.datetime.now().timestamp()
     print('Starting to unzip : /tmp/' + data['name'] + '...')
-    unZip('tmp', data['name'])
-    print('Uncompression finished with success. It took : ' + str(datetime.datetime.now().timestamp() - t) + ' seconds')
+    helper.unzip('tmp', data['name'])
+    print('Uncompress finished with success. It took : ' + str(datetime.datetime.now().timestamp() - t) + ' seconds')
 
-    # Clearing the useless files from the folder (like .zip, LICENCE, .gitattributes, README.md, etc...) [WE EXCLUDE .JSON FILES FROM DELETION]
+    # Clearing the useless files from the folder (like .zip, LICENCE, .gitattributes, README.md, etc...) [WE EXCLUDE
+    # .JSON FILES FROM DELETION]
     print('Deleting useless files ...')
-    clearDirectory('tmp')
+    helper.clear_directory('tmp')
 
     # Creating the .gma archive
     t = datetime.datetime.now().timestamp()
     print('Starting to create the .gma archive...')
-    baseName = getBaseDirectory()
-    createWorkshopArchive(gmad, baseName)
-    gmaName = baseName + '.gma'
-    print('The gma archive has been correctly created. It took : ' + str(
+    base_name = helper.get_base_directory()
+    create_ws_archive(base_name)
+    gma_name = base_name + '.gma'
+    print('The .gma archive has been correctly created. It took : ' + str(
         datetime.datetime.now().timestamp() - t) + ' seconds')
 
     # Updating the Workshop item
     t = datetime.datetime.now().timestamp()
     print('Starting to update the .gma archive...')
-    updateWorkshopItem(gmaName, fileid, data['body'])
+    update_ws_item(gma_name, steam.file_id, data['body'])
     print('The gma archive has been correctly uploaded to the Workshop. It took : ' + str(
         datetime.datetime.now().timestamp() - t) + ' seconds')
 
     return 0
 
 
-Main()
+# Script entry point
+if __name__ == '__main__':
+    main()
